chain: avoid locking the chain on duplicate pows
If a miner achieves duplicate pow, because the block is broadcast ahead
of time, all other miners attempting to mint a block containing the
winning tx end up in a lock being unable to broadcast a block contianing
the pending tx.

Instead of exploding on this condition, just skip over the operation
application and let the game roll on. They did the work, at least once.